src/python/requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
autobahn==17.6.2
Automat==0.6.0
beautifulsoup4==4.6.0
cachetools==2.0.1
certifi==2018.1.18
cffi==1.10.0
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==1.9
cycler==0.10.0
decorator==4.3.0
dill==0.2.7.1
dlib==19.10.0
Flask==1.0.2
Flask-SocketIO==3.0.1
future==0.16.0
gapic-google-cloud-logging-v2==0.91.3
gevent==1.2.2
google==1.9.3
google-api-core==0.1.4
google-auth==1.4.1
google-auth-httplib2==0.0.2
google-cloud==0.34.0
google-cloud-core==0.28.0
google-cloud-speech==0.35.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
greenlet==0.4.12
grpcio==1.9.1
httplib2==0.10.3
hyperlink==17.2.1
idna==2.6
imutils==0.4.6
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
jsocket==1.6
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
msgpack==0.5.6
msgpack-numpy==0.4.3.1
msgpack-python==0.4.8
networkx==2.1
numpy==1.14.0
oauth2client==3.0.0
opencv-python==3.4.0.12
pika==0.10.0
Pillow==5.1.0
ply==3.8
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-speech-v1==0.15.3
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycparser==2.18
pyOpenSSL==17.1.0
pyparsing==2.2.0
pysolr==3.6.0
python-dateutil==2.7.2
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
requests==2.18.4
rsa==3.4.2
scikit-image==0.13.1
scipy==1.0.1
service-identity==17.0.0
six==1.11.0
stt-watson==1.0.3
Twisted==17.5.0
txaio==2.8.0
urllib3==1.22
watson-developer-cloud==1.7.1
webrtcvad==2.0.10
websocket-client==0.44.0
Werkzeug==0.14.1
zope.interface==4.4.2
